requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-cloud-vision==3.5.0
Pillow==10.1.0
openpyxl==3.1.2
reportlab==4.0.7
requests==2.31.0
gunicorn==21.2.0
supabase==2.3.0
